Refactor packages name to landscapist
