+++ app/src/main/java/com/github/skydoves/landscapistdemo/theme/Shape.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.skydoves.fisherdemo.theme
+package com.github.skydoves.landscapistdemo.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
