+++ settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = "FisherDemo"
+rootProject.name = "LandscapistDemo"
 include ':app'
 include ':fresco'
